requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.0
piexif==1.1.3
Pillow==6.0.0
PyYAML==5.3
six==1.13.0
torch==1.3.1+cu92
torchvision==0.4.2+cu92
Werkzeug==0.16.0